Move import/export/share to toolbar menu
