--- android_app/app/src/main/res/menu/table_menu.xml
@@ -1,16 +1,0 @@
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.example.openscale.MainActivity">
-
-    <item
-        android:id="@+id/importData"
-        android:title="@string/label_import"/>
-
-    <item
-        android:id="@+id/exportData"
-        android:title="@string/label_export"/>
-
-    <item
-        android:id="@+id/shareData"
-        android:title="@string/label_share"/>
-</menu>
